refactor and add test cases
